updated with all todo

+++ projects/02_trivia_api/starter/backend/flaskr/__init__.py
@@ -20,7 +20,7 @@ def create_app(test_config=None):
     # create and configure the app
     app = Flask(__name__)
     setup_db(app)
-    CORS(app)
+    CORS(app, resources={"/": {"origins": "*"}})
 
       
     # after_request decorator to set Access-Control-Allow
@@ -30,20 +30,21 @@ def create_app(test_config=None):
         response.headers.add('Access-Control-Allow-Methods', 'GET, POST, PATCH, DELETE, OPTIONS')
         return response
 
-    # endpoint to handle GET requests for all available categories                
+    # endpoint to handle GET requests for all available categories  
+    # should return a list of all the categories and the total
+    # number of categories              
     @app.route('/categories', methods=["GET"])
     def get_specific_cat():
         all_cat = Category.query.order_by(Category.id).all()
         total_categories = len(all_cat)
         formatted_categories = {category.id:category.type for category in all_cat}
-
           
         if len(all_cat) == 0:
             abort(404)
           
         return jsonify({
             'success':True,
-            'categories': formatted_categories,
+            'categories':formatted_categories,
             "total_categories": total_categories
 
                     })
@@ -75,10 +76,12 @@ def create_app(test_config=None):
     
     
     # endpoint to DELETE question using a question ID.                
-    @app.route('/questions/<int:id>', methods=["DELETE"])
+    @app.route('/questions/<int:id>', methods=['GET','DELETE'])
     def delete_question(id):
         try:
-            question = Question.query.filter(Question.id == id).first_or_404()
+            question = Question.query.filter(Question.id == id).one_or_none()
+            if question is None:
+                abort(404)
             question.delete()
             selection = Question.query.order_by(Question.id).all()
             total_questions = len(selection)
@@ -92,7 +95,7 @@ def create_app(test_config=None):
         except:
             abort(422)
      
-    
+    # endpoint to create a new question
     @app.route('/questions/create', methods=["POST"])
     def create_question():
         body = request.get_json()
@@ -128,7 +131,6 @@ def create_app(test_config=None):
     # returns all the questions and the total number of questions 
     # for a specific category
 
-
     @app.route('/categories/<int:id>/questions')
 
     def search_question_cat(id):
@@ -140,7 +142,7 @@ def create_app(test_config=None):
         
 
         if len(selection) == 0:
-            return abort(404)
+            abort(404)
 
         try:
             current_questions = paginate_questions(request, selection)
@@ -149,8 +151,8 @@ def create_app(test_config=None):
                     'success':True,
                     "status_code":200,
                     'questions': current_questions,
-                    "totalQuestions": total_questions,
-                    "currentCategory": current_category
+                    "total_questions": total_questions,
+                    "current_category": current_category
                     })
         except:
             abort(422)
@@ -161,22 +163,17 @@ def create_app(test_config=None):
     def search_question():
 
         body = request.get_json()
-
         search_term = body.get('searchTerm') 
-        print(search_term)
-        
 
-        if not search_term:
-            abort(422)
+        selection = Question.query.filter(Question.question.ilike(f'%{search_term}%')).all()
+
+        
+        if len(selection) == 0:
+            abort(404)
 
         try:
-            selection = Question.query.filter(Question.question.ilike(f'%{search_term}%')).all()
-            total_questions = len(selection)
-            
-
-            if len(selection) == 0:
-                return abort(404)
             current_questions = paginate_questions(request, selection)
+            total_questions = len(selection)
             
             return jsonify({
                     'success':True,
@@ -198,7 +195,7 @@ def create_app(test_config=None):
 
 
         if body is None or quiz_category is None:
-            abort(404)
+            abort(400)
 
         # if All categories are selected that get all the questions
         # else, when a specific category is selected, select only the 
@@ -230,7 +227,8 @@ def create_app(test_config=None):
 
        
 
-        
+    # status codes and error messages
+    #     
     @app.errorhandler(404)
     def not_found(error):
         return jsonify({
@@ -253,7 +251,7 @@ def create_app(test_config=None):
         return jsonify({
           "success": False,
           "error": 400,
-          "message": "Bad Request"
+          "message": "bad request"
           }), 400
 
     @app.errorhandler(405)
@@ -261,91 +259,24 @@ def create_app(test_config=None):
         return jsonify({
           "success": False,
           "error": 405,
-          "message": "Method Not Allowed"
+          "message": "method not allowed"
           }), 405
+
+    @app.errorhandler(500)
+    def internal_server(error):
+        return jsonify({
+        'success': False,
+        'error': 500, 
+        'message': '''Oops, Somethis went wrong. The server ecnountered 
+                    an internal error or misconfiguration \n and was unable 
+                    to process your request. \n Please try again later.'''
+            }), 500
 
     return app
 
 
 
 
-'''
-@TODO: 
-.
-'''
-
-
-
-'''
-@TODO: 
-Create an endpoint to handle GET requests for questions, 
-including pagination (every 10 questions). 
-This endpoint should return a list of questions, 
-number of total questions, current category, categories. 
-
-TEST: At this point, when you start the application
-you should see questions and categories generated,
-ten questions per page and pagination at the bottom of the screen for three pages.
-Clicking on the page numbers should update the questions. 
-'''
-
-'''
-@TODO: 
-Create an endpoint to DELETE question using a question ID. 
-
-TEST: When you click the trash icon next to a question, the question will be removed.
-This removal will persist in the database and when you refresh the page. 
-'''
-
-'''
-@TODO: 
-Create an endpoint to POST a new question, 
-which will require the question and answer text, 
-category, and difficulty score.
-
-TEST: When you submit a question on the "Add" tab, 
-the form will clear and the question will appear at the end of the last page
-of the questions list in the "List" tab.  
-'''
-
-'''
-@TODO: 
-Create a POST endpoint to get questions based on a search term. 
-It should return any questions for whom the search term 
-is a substring of the question. 
-
-TEST: Search by any phrase. The questions list will update to include 
-only question that include that string within their question. 
-Try using the word "title" to start. 
-'''
-
-'''
-@TODO: 
-Create a GET endpoint to get questions based on category. 
-
-TEST: In the "List" tab / main screen, clicking on one of the 
-categories in the left column will cause only questions of that 
-category to be shown. 
-'''
-
-
-'''
-@TODO: 
-Create a POST endpoint to get questions to play the quiz. 
-This endpoint should take category and previous question parameters 
-and return a random questions within the given category, 
-if provided, and that is not one of the previous questions. 
-
-TEST: In the "Play" tab, after a user selects "All" or a category,
-one question at a time is displayed, the user is allowed to answer
-and shown whether they were correct or not. 
-'''
-
-'''
-@TODO: 
-Create error handlers for all expected errors 
-including 404 and 422. 
-'''
